Fixed various problems in virtual display code and added more tests

+++ unixonly.py
@@ -38,27 +38,53 @@ class VirtualDisplayFinder:
     def getDisplay(self):
         if len(self.machines) == 0:
             return
+
+        usableDisplay, emptyMachine = self.classifyMachines()
+        if usableDisplay:
+            return usableDisplay
+        elif emptyMachine:
+            return self.startServer(emptyMachine)
+        else:
+            print "WARNING - virtual display test command failed on all machines - attempting to use first one anyway!"
+            return self.machines[0] + ":42.0"
+    def classifyMachines(self):
+        emptyMachine = None
+        # Try to find an existing display we can connect to
         for machine in self.machines:
             self.diag.info("Looking for virtual display on " + machine)
-            display = self.findDisplay(machine)
-            if display and display != "":
-                return display
-            if display != None:
-                self.killServer(machine)
-        for machine in self.machines:
-            display = self.startServer(machine)
-            if display and display != "":
-                return display
+            display, displayErrs = self.findDisplay(machine)
+            if display:
+                connErrors = self.getConnectionErrors(display)
+                if len(connErrors) > 0:
+                    print "Failed to connect to virtual server on", machine + "\n" + connErrors
+                    if self.killServer(machine) and not emptyMachine:
+                        emptyMachine = machine
+                else:
+                    return display, emptyMachine
+            else:
+                if displayErrs:
+                    print "Could not get display info from", machine + ":\n" + displayErrs
+                elif not emptyMachine:
+                    emptyMachine = machine
+        return None, emptyMachine
     def killServer(self, server):
         # Xvfb servers get overloaded after a while. If they do, kill them
-        line = self.getPsOutput(server, ownProcesses=1)
+        line, errors = self.getPsOutput(server)
         if len(line) == 0:
-            # We will only kill servers that were started by TextTest (have number 42!)
-            return
-        # On Linux fourth column of ps output is pid
-        pidStr = line.split()[3]
-        print "Trying to kill unusable Xvfb process", pidStr, "on machine " + server
+            # If it's been killed in the meantime, all is well...
+            return True
+
+        words = line.split()
+        # Assumes linux ps output (!)
+        processOwner = words[2]
+        if processOwner != os.getenv("USER"):
+            print "Unusable Xvfb process on machine " + server + " owned by user " + processOwner
+            return False
+    
+        pidStr = words[3]
+        print "Killing unusable Xvfb process", pidStr, "on machine " + server
         os.system(self.getSysCommand(server, "kill -9 " + pidStr, background=0))
+        return True
     def getSysCommand(self, server, command, background=1):
         if server == gethostname():
             if background:
@@ -82,31 +108,27 @@ class VirtualDisplayFinder:
         os.system(self.getSysCommand(server, "xclock -display " + serverName))
         os.system(self.getSysCommand(server, "xterm -display " + serverName + " -e xhost +"))
         return serverName
-    def getPsOutput(self, server, ownProcesses):
+    def getPsOutput(self, server):
         self.diag.info("Getting ps output from " + server)
         psCommand = self.getSysCommand(server, "ps -efl | grep Xvfb | grep -v grep", background=0)
-        lines = os.popen(psCommand).readlines()
-        for line in lines:
-            if line.find("Xvfb") != -1 and (not ownProcesses or line.find("42") != -1):
-                return line
-        return ""
+        cin, cout,cerr = os.popen3(psCommand)
+        for line in cout.readlines():
+            if line.find("Xvfb") != -1 and line.find("42") != -1:
+                return line, ""
+        return "", cerr.read()
     def findDisplay(self, server):
-        line = self.getPsOutput(server, ownProcesses=0)
+        line, errors = self.getPsOutput(server)
         if len(line):
-            self.diag.info("Found Xvfb process running:" + "\n" + line)
+            self.diag.info("Found Xvfb process running:\n" + line)
             serverName = server + line.split()[-1] + ".0"
-            testCommandLine = self.getSysCommand(gethostname(), "xterm -display " + serverName + " -e echo test", background=0)
-            self.diag.info("Testing with command '" + testCommandLine + "'")
-            (cin, cout, cerr) = os.popen3(testCommandLine)
-            lines = cerr.readlines()
-            if len(lines) == 0:
-                return serverName
-            else:
-                print "Failed to connect to virtual server on", server
-                for line in lines:
-                    print line.strip()
-                return ""
-        return None
+            return serverName, ""
+        else:
+            return "", errors
+    def getConnectionErrors(self, serverName):
+        testCommandLine = self.getSysCommand(gethostname(), "xterm -display " + serverName + " -e echo test", background=0)
+        self.diag.info("Testing with command '" + testCommandLine + "'")
+        (cin, cout, cerr) = os.popen3(testCommandLine)
+        return cerr.read()
    
 def isCompressed(path):
     if os.path.getsize(path) == 0:
